requirements.txt
accelerate==0.23.0
addict==2.4.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
blis==0.7.11
brotlipy==0.7.0
catalogue==2.0.10
certifi @ file:///croot/certifi_1690232220950/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpathlib==0.15.1
confection==0.1.3
contourpy==1.1.1
cryptography @ file:///croot/cryptography_1694444244250/work
cycler==0.12.1
cymem==2.0.8
datasets==2.14.5
dill==0.3.7
emoji==2.8.0
en-core-web-lg @ file:///root/autodl-tmp/fqy/models/en_core_web_lg-3.7.0.tar.gz#sha256=996b7c74422139dcb4acf5b640f2836af1611e38b7728e3303b8c768cb8555a4
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
evaluate==0.4.0
filelock==3.12.4
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.17.3
idna @ file:///croot/idna_1666125576474/work
importlib-metadata==6.8.0
importlib-resources==6.1.0
install==1.3.5
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.0.3
pathy==0.10.2
Pillow @ file:///croot/pillow_1696580024257/work
platformdirs==3.11.0
preshed==3.0.9
protobuf==4.24.4
psutil==5.9.5
pyarrow==13.0.0
pycocotools==2.0.7
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests @ file:///croot/requests_1690400202158/work
responses==0.18.0
rich==13.6.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
smart-open==6.4.0
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stanza==1.6.1
supervision==0.16.0
thinc==8.2.1
threadpoolctl==3.2.0
timm==0.9.8
tokenizers==0.14.1
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.66.1
transformers==4.34.0
typer==0.9.0
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1686163155763/work
wasabi==1.1.2
weasel==0.3.2
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.2
zipp==3.17.0
